+++ src/FileManager.php
@@ -4,13 +4,15 @@ declare(strict_types=1);
 
 namespace Moudarir\FileManager;
 
+use Moudarir\FileManager\Collections\CollectionBuilder;
+use Moudarir\FileManager\Collections\CollectionInterface;
+use Moudarir\FileManager\Config\FileManagerConfig;
 use Moudarir\FileManager\Exceptions\FileManagerException;
 use Moudarir\FileManager\Image\ImageConverter;
 use Moudarir\FileManager\Image\ImageCropper;
 use Moudarir\FileManager\Image\ImageResizer;
 use Moudarir\FileManager\Image\ImageWatermarker;
 use Moudarir\FileManager\Upload\FileUploader;
-use Moudarir\FileManager\Upload\UploadedFileCollection;
 
 final class FileManager
 {
@@ -29,6 +31,8 @@ final class FileManager
 
     private bool $watermarkRequested = false;
 
+    private ?CollectionInterface $collection = null;
+
     public function __construct(private readonly FileManagerConfig $config)
     {
     }
@@ -36,46 +40,62 @@ final class FileManager
     /**
      * @throws FileManagerException
      */
-    public function files(): UploadedFileCollection
+    public function files(): CollectionInterface
     {
-        if ($this->uploadRequested === false) {
-            throw FileManagerException::uploadRequestMandatory();
+        if (isset($this->collection) === false) {
+            if ($this->uploadRequested === false) {
+                throw FileManagerException::uploadRequestMandatory();
+            }
+
+            $this->collection = new FileUploader($this->config->uploadConfig())
+                ->upload($this->filepath);
         }
 
-        $collection = new FileUploader($this->config->uploadConfig())
-            ->upload($this->filepath);
-
-        if ($collection->isEmpty()) {
-            return $collection;
+        if ($this->collection->isEmpty()) {
+            return $this->collection;
         }
 
         if ($this->cropRequested === true && empty($this->croppingConfig) === false) {
             ImageCropper::create(
-                $collection,
+                $this->collection,
                 $this->config->imageCropConfig(),
                 $this->croppingConfig
             )->crop();
         }
 
         if ($this->resizeRequested === true) {
-            ImageResizer::create($collection, $this->config->imageResizeConfig());
+            ImageResizer::create($this->collection, $this->config->imageResizeConfig());
         }
 
         if ($this->watermarkRequested === true) {
-            ImageWatermarker::create($collection, $this->config->imageWatermarkConfig());
+            ImageWatermarker::create($this->collection, $this->config->imageWatermarkConfig());
         }
 
         if ($this->convertRequested === true) {
-            ImageConverter::create($collection, $this->config->imageConvertConfig());
+            ImageConverter::create($this->collection, $this->config->imageConvertConfig());
         }
 
-        return $collection;
+        return $this->collection;
     }
 
     public function upload(?string $filepath = null): self
     {
         $this->uploadRequested = true;
         $this->filepath = $filepath;
+
+        return $this;
+    }
+
+    /**
+     * @throws FileManagerException
+     */
+    public function buildUploadedFileCollection(array $files, bool $ignoreInvalidFiles = false): self
+    {
+        $this->collection = CollectionBuilder::create(
+            $files,
+            $this->config->collectionBuilderConfig(),
+            $ignoreInvalidFiles
+        );
 
         return $this;
     }